+++ src/mocks/MockERC20.sol
@@ -19,11 +19,11 @@ contract MockERC20 is ERC20 {
         return _decimals;
     }
 
-    function mint(address to, uint amount) external {
+    function mint(address to, uint256 amount) external {
         _mint(to, amount);
     }
 
-    function burn(address from, uint amount) external {
+    function burn(address from, uint256 amount) external {
         _burn(from, amount);
     }
 }
