+++ src/mocks/MockEarlySellPenaltyHook.sol
@@ -10,34 +10,30 @@ import "../interfaces/IEarlySellPenaltyHook.sol";
  */
 contract MockEarlySellPenaltyHook is IEarlySellPenaltyHook {
     // Mock state variables
-    mapping(address => uint) private mockBuyerTimestamps;
-    uint public penaltyDeclineRatePerHour = 10; // Default 1% per hour
-    uint public maxPenaltyDurationHours = 100;
+    mapping(address => uint256) private mockBuyerTimestamps;
+    uint256 public penaltyDeclineRatePerHour = 10; // Default 1% per hour
+    uint256 public maxPenaltyDurationHours = 100;
     bool public penaltyActive = true;
 
     // Call tracking for tests
-    uint public buyCallCount;
-    uint public sellCallCount;
+    uint256 public buyCallCount;
+    uint256 public sellCallCount;
     address public lastBuyer;
     address public lastSeller;
-    uint public lastBaseBondingToken;
-    uint public lastBaseInputToken;
-    uint public lastPenaltyApplied;
+    uint256 public lastBaseBondingToken;
+    uint256 public lastBaseInputToken;
+    uint256 public lastPenaltyApplied;
 
     // Override behavior controls
     bool public shouldFailOnBuy;
     bool public shouldFailOnSell;
-    uint public mockCurrentTime;
+    uint256 public mockCurrentTime;
     bool public useRealBlockTimestamp = true;
 
-    function buy(
-        address buyer,
-        uint baseBondingToken,
-        uint baseInputToken
-    )
+    function buy(address buyer, uint256 baseBondingToken, uint256 baseInputToken)
         external
         override
-        returns (uint fee, int deltaBondingToken)
+        returns (uint256 fee, int256 deltaBondingToken)
     {
         if (shouldFailOnBuy) {
             revert("MockEarlySellPenaltyHook: Buy failed");
@@ -49,7 +45,7 @@ contract MockEarlySellPenaltyHook is IEarlySellPenaltyHook {
         lastBaseInputToken = baseInputToken;
 
         // Record timestamp for buyer
-        uint currentTime = useRealBlockTimestamp ? block.timestamp : mockCurrentTime;
+        uint256 currentTime = useRealBlockTimestamp ? block.timestamp : mockCurrentTime;
         mockBuyerTimestamps[buyer] = currentTime;
 
         emit BuyerTimestampRecorded(buyer, currentTime);
@@ -58,14 +54,10 @@ contract MockEarlySellPenaltyHook is IEarlySellPenaltyHook {
         return (0, 0);
     }
 
-    function sell(
-        address seller,
-        uint baseBondingToken,
-        uint baseInputToken
-    )
+    function sell(address seller, uint256 baseBondingToken, uint256 baseInputToken)
         external
         override
-        returns (uint fee, int deltaBondingToken)
+        returns (uint256 fee, int256 deltaBondingToken)
     {
         if (shouldFailOnSell) {
             revert("MockEarlySellPenaltyHook: Sell failed");
@@ -77,11 +69,11 @@ contract MockEarlySellPenaltyHook is IEarlySellPenaltyHook {
         lastBaseInputToken = baseInputToken;
 
         // Calculate penalty fee
-        uint penaltyFee = calculatePenaltyFee(seller);
+        uint256 penaltyFee = calculatePenaltyFee(seller);
         lastPenaltyApplied = penaltyFee;
 
         if (penaltyFee > 0) {
-            uint hoursElapsed = getHoursElapsed(seller);
+            uint256 hoursElapsed = getHoursElapsed(seller);
             emit PenaltyApplied(seller, penaltyFee, hoursElapsed);
         }
 
@@ -89,11 +81,11 @@ contract MockEarlySellPenaltyHook is IEarlySellPenaltyHook {
         return (penaltyFee, 0);
     }
 
-    function getBuyerTimestamp(address buyer) external view override returns (uint timestamp) {
+    function getBuyerTimestamp(address buyer) external view override returns (uint256 timestamp) {
         return mockBuyerTimestamps[buyer];
     }
 
-    function setPenaltyParameters(uint _declineRatePerHour, uint _maxDurationHours) external override {
+    function setPenaltyParameters(uint256 _declineRatePerHour, uint256 _maxDurationHours) external override {
         penaltyDeclineRatePerHour = _declineRatePerHour;
         maxPenaltyDurationHours = _maxDurationHours;
         emit PenaltyParametersUpdated(_declineRatePerHour, _maxDurationHours);
@@ -104,44 +96,49 @@ contract MockEarlySellPenaltyHook is IEarlySellPenaltyHook {
         emit PenaltyStatusChanged(_active);
     }
 
-    function getPenaltyParameters() external view override returns (uint declineRate, uint maxDuration, bool active) {
+    function getPenaltyParameters()
+        external
+        view
+        override
+        returns (uint256 declineRate, uint256 maxDuration, bool active)
+    {
         return (penaltyDeclineRatePerHour, maxPenaltyDurationHours, penaltyActive);
     }
 
-    function calculatePenaltyFee(address seller) public view override returns (uint penaltyFee) {
+    function calculatePenaltyFee(address seller) public view override returns (uint256 penaltyFee) {
         if (!penaltyActive) return 0;
 
-        uint lastBuyTimestamp = mockBuyerTimestamps[seller];
+        uint256 lastBuyTimestamp = mockBuyerTimestamps[seller];
         if (lastBuyTimestamp == 0) {
             // First-time seller without previous buy gets maximum penalty
             return 1000;
         }
 
-        uint currentTime = useRealBlockTimestamp ? block.timestamp : mockCurrentTime;
-        uint timeElapsed = currentTime - lastBuyTimestamp;
-        uint hoursElapsed = timeElapsed / 3600;
+        uint256 currentTime = useRealBlockTimestamp ? block.timestamp : mockCurrentTime;
+        uint256 timeElapsed = currentTime - lastBuyTimestamp;
+        uint256 hoursElapsed = timeElapsed / 3600;
 
         if (hoursElapsed >= maxPenaltyDurationHours) {
             return 0;
         }
 
         // Calculate declining penalty: starts at 100% (1000), declines by penaltyDeclineRatePerHour per hour
-        uint penalty = 1000 - (hoursElapsed * penaltyDeclineRatePerHour);
+        uint256 penalty = 1000 - (hoursElapsed * penaltyDeclineRatePerHour);
 
         // Ensure penalty doesn't go negative
         return penalty > 1000 ? 0 : penalty;
     }
 
     // Helper functions for testing
-    function getHoursElapsed(address seller) public view returns (uint) {
-        uint lastBuyTimestamp = mockBuyerTimestamps[seller];
+    function getHoursElapsed(address seller) public view returns (uint256) {
+        uint256 lastBuyTimestamp = mockBuyerTimestamps[seller];
         if (lastBuyTimestamp == 0) return 0;
 
-        uint currentTime = useRealBlockTimestamp ? block.timestamp : mockCurrentTime;
+        uint256 currentTime = useRealBlockTimestamp ? block.timestamp : mockCurrentTime;
         return (currentTime - lastBuyTimestamp) / 3600;
     }
 
-    function setMockTimestamp(uint timestamp) external {
+    function setMockTimestamp(uint256 timestamp) external {
         mockCurrentTime = timestamp;
         useRealBlockTimestamp = false;
     }
@@ -150,7 +147,7 @@ contract MockEarlySellPenaltyHook is IEarlySellPenaltyHook {
         useRealBlockTimestamp = true;
     }
 
-    function setBuyerTimestamp(address buyer, uint timestamp) external {
+    function setBuyerTimestamp(address buyer, uint256 timestamp) external {
         mockBuyerTimestamps[buyer] = timestamp;
     }
